requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
notion-database==1.1.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.18
Werkzeug==3.0.1
